requirements.txt
Pillow>=10.0.0
pillow-heif>=0.13.0
img2pdf>=0.4.4